Убрал диагностику линкера о записи вне массива

+++ inc/formats.h
@@ -19,8 +19,8 @@
 extern "C" {
 #endif /* __cplusplus */
 
-uint_fast8_t local_snprintf_P( char * __restrict buffer, uint_fast8_t count, const char * __restrict format, ... ) __attribute__ ((__format__ (__printf__, 3, 4)));
-uint_fast8_t local_vsnprintf_P( char * __restrict buffer, uint_fast8_t count, const char * __restrict format, va_list ap ) __attribute__ ((__format__ (__printf__, 3, 0)));
+int local_snprintf_P( char * __restrict buffer, size_t count, const char * __restrict format, ... ) __attribute__ ((__format__ (__printf__, 3, 4)));
+int local_vsnprintf_P( char * __restrict buffer, size_t count, const char * __restrict format, va_list ap ) __attribute__ ((__format__ (__printf__, 3, 0)));
 // Отладочная печать
 void debug_printf_P(const char * __restrict format, ... ) __attribute__ ((__format__ (__printf__, 1, 2)));
